pyproject.toml
[project]
name = "car-dealership-bot"
version = "2.0.0"
description = "Car Dealership Telegram Chatbot"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.111,<0.112",
    "uvicorn[standard]>=0.30,<0.31",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29,<0.30",
    "alembic>=1.13,<2.0",
    "langchain>=0.3,<0.4",
    "langchain-openai>=0.2,<0.3",
    "langchain-core>=0.3,<0.4",
    "langgraph>=0.2,<0.3",
    "langgraph-checkpoint-postgres>=1.0,<2.0",
    "psycopg[binary]>=3.1,<4.0",
    "openai>=1.30,<2.0",
    "langfuse>=2.36,<3.0",
    "google-api-python-client>=2.130,<3.0",
    "google-auth-httplib2>=0.2,<0.3",
    "google-auth-oauthlib>=1.2,<2.0",
    "python-telegram-bot>=21.3,<22.0",
    "httpx>=0.27,<0.28",
    "python-json-logger>=2.0,<3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2,<9.0",
    "pytest-asyncio>=0.23,<0.24",
    "pytest-mock>=3.14,<4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
